general : debugging

+++ src/main/scala/com/muhalfian/spark/jobs/Dictionary.scala
@@ -71,7 +71,7 @@ object Dictionary extends StreamUtils {
     val selectedDF = preprocessDF.select("link", "source", "description", "image", "publish_date", "title", "text", "text_preprocess")
                         .withColumn("text_selected", TextTools.select(col("text_preprocess")))
 
-    println(masterWord.count)
+    println(masterWord.count.toInt)
     masterWord.show()
 
     // ======================== SAVE DICTIONARY ================================
